logmsg/nvtable: improve nv_table_get_next_size() for incremental growth
Whenever NVTable needs to grow (either during clone or realloc), the new
nv_table_get_next_size() function calculates the right new size, while
ensuring some "additional_space" becomes available in the new clone.

Previously, we always erred on the side of increasing the allocation size,
so that any expected new changes would not need another bump of the size.
This algorithm however, can waste a lot of memory.

The new algorithm is:
  * increase at least 256 bytes,
  * only increase the size if the current allocation does not fit
  * use 4k aligned sizes for large NVTables (>4kB) or 1k for small ones

We might have slightly more reallocs this way, but the total allocation of a
LogMessage should be a lot less.

+++ lib/logmsg/nvtable.h
@@ -313,7 +313,7 @@ NVTable *nv_table_new(gint num_static_values, gint index_size_hint, gint init_le
 NVTable *nv_table_init_borrowed(gpointer space, gsize space_len, gint num_static_entries);
 gboolean nv_table_realloc(NVTable *self, NVTable **new_nv_table);
 NVTable *nv_table_compact(NVTable *self);
-NVTable *nv_table_clone(NVTable *self, gint additional_space);
+NVTable *nv_table_clone(NVTable *self, gsize additional_space);
 NVTable *nv_table_ref(NVTable *self);
 void nv_table_unref(NVTable *self);
